ajax_auth/fastapi_backends/oauth2_google/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
crypto==1.4.1
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.115.12
google-auth==2.40.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jose==1.0.0
mysql-connector-python==9.3.0
Naked==0.1.32
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-jose==3.5.0
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
